refactor: move home content out from header
